# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready Go Fiber boilerplate with RBAC, Redis caching, and PostgreSQL.",
        "title": "Go Fiber Boilerplate API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/api/v1",
    "paths": {
...
        "/patients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all patients with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get all patients",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID filter",
                        "name": "room_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Condition level filter",
                        "name": "condition_level",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Admission status filter",
                        "name": "is_admitted",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/docs.PatientDocResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/docs.MetaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admits a new patient to the hospital",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Admit a new patient",
                "parameters": [
                    {
                        "description": "Patient data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_patient.AdmitPatientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.PatientDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a patient by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get a patient by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.PatientDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a patient by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patient update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_patient.UpdatePatientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/docs.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/docs.PatientDocResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a patient by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Delete a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{uuid}/condition": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a patient's condition level",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Update patient condition level",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Condition update data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_patient.UpdateConditionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{uuid}/discharge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Discharges a patient from the hospital",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Discharge a patient",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/docs.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/docs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.ErrorResponse": {
            "description": "Standard error response wrapper",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error_code": {
                    "type": "integer",
                    "example": -1001
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error occurred"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "docs.MetaResponse": {
            "description": "Pagination metadata for list responses",
            "type": "object",
            "properties": {
                "is_back": {
                    "type": "boolean",
                    "example": false
                },
                "is_next": {
                    "type": "boolean",
                    "example": true
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "docs.PatientDocResponse": {
            "description": "Patient information",
            "type": "object",
            "properties": {
                "admission_date": {
                    "type": "string"
                },
                "condition_level": {
                    "type": "string",
                    "example": "stable"
                },
                "created_at": {
                    "type": "string"
                },
                "diagnosis": {
                    "type": "string",
                    "example": "Pneumonia"
                },
                "discharge_date": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "example": "female"
                },
                "medical_record_number": {
                    "type": "string",
                    "example": "MRN-001"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "room_name": {
                    "type": "string",
                    "example": "Room 101"
                },
                "room_type": {
                    "type": "string",
                    "example": "patient_room"
                },
                "room_uuid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "docs.SuccessResponse": {
            "description": "Standard success response wrapper",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "meta": {
                    "$ref": "#/definitions/docs.MetaResponse"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_module_patient.AdmitPatientRequest": {
            "type": "object",
            "required": [
                "medical_record_number",
                "name"
            ],
            "properties": {
                "condition_level": {
                    "enum": [
                        "critical",
                        "serious",
                        "moderate",
                        "stable",
                        "good"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_patient.ConditionLevel"
                        }
                    ]
                },
                "date_of_birth": {
                    "type": "string"
                },
                "diagnosis": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "medical_record_number": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "notes": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
        "internal_module_patient.ConditionLevel": {
            "type": "string",
            "enum": [
                "critical",
                "serious",
                "moderate",
                "stable",
                "good"
            ],
            "x-enum-varnames": [
                "ConditionCritical",
                "ConditionSerious",
                "ConditionModerate",
                "ConditionStable",
                "ConditionGood"
            ]
        },
        "internal_module_patient.UpdateConditionRequest": {
            "type": "object",
            "required": [
                "condition_level"
            ],
            "properties": {
                "condition_level": {
                    "enum": [
                        "critical",
                        "serious",
                        "moderate",
                        "stable",
                        "good"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_patient.ConditionLevel"
                        }
                    ]
                }
            }
        },
        "internal_module_patient.UpdatePatientRequest": {
            "type": "object",
            "properties": {
                "condition_level": {
                    "enum": [
                        "critical",
                        "serious",
                        "moderate",
                        "stable",
                        "good"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_module_patient.ConditionLevel"
                        }
                    ]
                },
                "diagnosis": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "notes": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}. Get token from /auth/login endpoint.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}